+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginTest.kt
@@ -2,8 +2,8 @@ package io.sentry.android.gradle
 
 import java.io.File
 import kotlin.test.assertNotEquals
-// import kotlin.test.assertNotNull
-// import kotlin.test.assertNull
+import kotlin.test.assertNotNull
+import kotlin.test.assertNull
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.internal.PluginUnderTestMetadataReading
 import org.junit.Before
@@ -127,64 +127,53 @@ class SentryPluginTest(
         verifyProguardUuid(testProjectDir.root)
     }
 
-//    @Test
-//    fun `does not include a UUID in the APK`() {
-//        // isMinifyEnabled is disabled by default in debug builds
-//        runner
-//            .appendArguments(":app:assembleDebug")
-//            .build()
-//
-//        verifyNoProguardUuid(testProjectDir.root)
-//    }
+    @Test
+    fun `does not include a UUID in the APK`() {
+        // isMinifyEnabled is disabled by default in debug builds
+        runner
+            .appendArguments(":app:assembleDebug")
+            .build()
 
-//    @Test
-//    fun `creates uploadNativeSymbolsForRelease task if uploadNativeSymbols is enabled and non debuggable app`() {
-//        applyUploadNativeSymbols()
-//
-//        val build = runner
-//            .appendArguments(":app:assembleRelease")
-//            .build()
-//
-//        assertNotNull(build.task("uploadNativeSymbolsForRelease"))
-//    }
+        verifyNoProguardUuid(testProjectDir.root)
+    }
 
-//    @Test
-//    fun `does not create uploadNativeSymbolsForRelease task if non debuggable app`() {
-//        applyUploadNativeSymbols()
-//
-//        val build = runner
-//            .appendArguments(":app:assembleDebug")
-//            .build()
-//
-//        assertNull(build.task("uploadNativeSymbolsForDebug"))
-//    }
+    @Test
+    fun `creates uploadNativeSymbolsForRelease task if uploadNativeSymbols is enabled`() {
+        applyUploadNativeSymbols()
 
-//    @Test
-//    fun `does not create uploadNativeSymbolsForRelease task if uploadNativeSymbols is disabled`() {
-//        applyUploadNativeSymbols(false)
-//
-//        val build = runner
-//            .appendArguments(":app:assembleRelease")
-//            .build()
-//
-//        assertNull(build.task("uploadNativeSymbolsForRelease"))
-//    }
+        val build = runner
+            .appendArguments(":app:assembleRelease")
+            .build()
 
-//    private fun applyUploadNativeSymbols(uploadNativeSymbols: Boolean = true) {
-//        appBuildFile.writeText(
-//            // language=Groovy
-//            """
-//                plugins {
-//                  id "com.android.application"
-//                  id "io.sentry.android.gradle"
-//                }
-//
-//                sentry {
-//                  uploadNativeSymbols = $uploadNativeSymbols
-//                }
-//            """.trimIndent()
-//        )
-//    }
+        assertNotNull(build.task("uploadNativeSymbolsForRelease"))
+    }
+
+    @Test
+    fun `does not create uploadNativeSymbolsForRelease task if non debuggable app`() {
+        applyUploadNativeSymbols()
+
+        val build = runner
+            .appendArguments(":app:assembleDebug")
+            .build()
+
+        assertNull(build.task("uploadNativeSymbolsForDebug"))
+    }
+
+    private fun applyUploadNativeSymbols() {
+        appBuildFile.writeText(
+            // language=Groovy
+            """
+                plugins {
+                  id "com.android.application"
+                  id "io.sentry.android.gradle"
+                }
+
+                sentry {
+                  uploadNativeSymbols = true
+                }
+            """.trimIndent()
+        )
+    }
 
     companion object {